Fix unidoc classpath problems
Previously, unidoc would nondeterministically fail because we had conflicting
versions of com.google.protobuf.Internal on its classpath, and its classpath
was not ordered.

This commit updates sbt-unidoc to version 0.4.3 which makes the classpath
ordering deterministic (https://github.com/sbt/sbt-unidoc/pull/64) and excludes
the older version of protobuf so the unidocs are correctly generated with the
newer version.